Refactor homepage
This commit introduces a new version of my homepage.

I was tired of messing around with bootstrap and npm, but the main
driver behind this decision was not liking having to scroll through 3 or
4 sections to see such a short amount of content.

Remembered http://motherfuckingwebsite.com and just went with a similar
approach, which fits what I currently do a lot better.

It might not be as pretty, but it certainly is much more readable (and
probably more original also)